+++ fj-doc-maven-plugin/src/main/java/org/fugerit/java/doc/project/facade/FlavourFacade.java
@@ -6,7 +6,6 @@ import freemarker.template.TemplateException;
 import lombok.extern.slf4j.Slf4j;
 import org.fugerit.java.core.cfg.ConfigRuntimeException;
 import org.fugerit.java.core.function.SafeFunction;
-import org.fugerit.java.core.io.StreamIO;
 import org.fugerit.java.core.lang.helpers.ClassHelper;
 import org.fugerit.java.core.lang.helpers.StringUtils;
 import org.fugerit.java.core.lang.helpers.reflect.MethodHelper;
@@ -19,7 +18,6 @@ import org.fugerit.java.doc.project.facade.flavour.extra.ParamConfig;
 
 import java.io.*;
 import java.lang.reflect.Field;
-import java.lang.reflect.Method;
 import java.util.*;
 
 @Slf4j
@@ -28,6 +26,8 @@ public class FlavourFacade {
     private FlavourFacade() {}
 
     public static final String FLAVOUR_VANILLA = "vanilla";
+
+    public static final String FLAVOUR_DIRECT = "direct";
 
     public static final String FLAVOUR_QUARKUS_3 = "quarkus-3";
 
@@ -50,7 +50,7 @@ public class FlavourFacade {
     private static final Properties FLAVOURS_DEFAULT_VERSION = PropsIO.loadFromClassLoaderSafe( "config/flavour/flavour_versions_default.properties" );
 
     public static final Set<String> SUPPORTED_FLAVOURS = Collections.unmodifiableSet(
-            new HashSet<>( Arrays.asList( FLAVOUR_VANILLA, FLAVOUR_QUARKUS_3, FLAVOUR_QUARKUS_3_GRADLE, FLAVOUR_QUARKUS_3_GRADLE_KTS,
+            new HashSet<>( Arrays.asList( FLAVOUR_VANILLA, FLAVOUR_DIRECT, FLAVOUR_QUARKUS_3, FLAVOUR_QUARKUS_3_GRADLE, FLAVOUR_QUARKUS_3_GRADLE_KTS,
                     FLAVOUR_QUARKUS_3_PROPERTIES, FLAVOUR_QUARKUS_2, FLAVOUR_MICRONAUT_4, FLAVOUR_SPRINGBOOT_3, FLAVOUR_OPENLIBERTY ) ) );
 
     public static boolean isGradleKtsFlavour(String flavour ) {
@@ -63,7 +63,7 @@ public class FlavourFacade {
         return prop;
     });
 
-    public static void initProject( FlavourContext context ) throws IOException, TemplateException {
+    public static String initProject( FlavourContext context ) throws IOException, TemplateException {
         log.info( "generate flavour : {}", context.getFlavour() );
         String actualFlavour = MAP_FLAVOURS.getProperty( context.getFlavour(), context.getFlavour() );
         if ( SUPPORTED_FLAVOURS.contains( actualFlavour ) ) {
@@ -72,6 +72,7 @@ public class FlavourFacade {
         } else {
             throw new ConfigRuntimeException( String.format( "flavour not supported : %s", context.getFlavour() ) );
         }
+        return actualFlavour;
     }
 
     public static void checkFlavour( FlavourContext context, String actualFlavour ) {
@@ -147,12 +148,7 @@ public class FlavourFacade {
 
     private static void initFlavour( FlavourContext context, String actualFlavour ) throws IOException, TemplateException {
         // copy all resources
-        String listFilePath = String.format( "config/flavour/%s-copy.txt", actualFlavour );
-        String baseFlavourPath = String.format( "config/flavour/%s/", actualFlavour );
-        log.info( "loading list file {}, base flavour path {}", listFilePath, baseFlavourPath );
-        try (BufferedReader reader = new BufferedReader( new InputStreamReader(ClassHelper.loadFromDefaultClassLoader( listFilePath ) ) ) )  {
-            reader.lines().forEach( s -> copyFlavourFile( s, context.getProjectFolder(), baseFlavourPath ) );
-        }
+        FeatureFacade.copyFlavourList( context.getProjectFolder(), actualFlavour );
         // freemarker resources
         Map<String, Object> data = new HashMap<>();
         data.put( "context", context );
@@ -167,31 +163,11 @@ public class FlavourFacade {
         }
     }
 
-    private static void insureParent( File file ) throws IOException {
-        File parentFile = file.getParentFile();
-        if ( !parentFile.exists() ) {
-            log.info( "creates parent directory {}, mkdirs:?", parentFile.getCanonicalPath(), parentFile.mkdirs() );
-        }
-    }
-
-    private static void copyFlavourFile( String path, File baseFolder, String basePath ) {
-        SafeFunction.apply( () -> {
-            File outputFile = new File( baseFolder, path );
-            insureParent( outputFile );
-            String fullPath = basePath+path;
-            log.info( "copy path '{}' to file '{}'", fullPath, outputFile.getCanonicalPath() );
-            try ( InputStream is = ClassHelper.loadFromDefaultClassLoader( fullPath );
-                    FileOutputStream os = new FileOutputStream( outputFile ) ) {
-                StreamIO.pipeStream( is, os, StreamIO.MODE_CLOSE_NONE );
-            }
-        } );
-    }
-
     private static void processEntry( ProcessEntry entry, Map<String, Object> data ) {
         log.info( "process entry : {}", entry );
         SafeFunction.apply( () -> {
             File toFile = new File( entry.getTo() );
-            insureParent( toFile );
+            FeatureFacade.insureParent( toFile );
             FreemarkerTemplateFacade.processFile( entry.getFrom(), toFile, data );
         } );
     }