+++ fj-doc-playground-quarkus/src/main/java/org/fugerit/java/doc/playground/init/ProjectInitInput.java
@@ -50,4 +50,12 @@ public class ProjectInitInput {
     @Setter
     private List<String> extensionList;
 
+    @Getter
+    @Setter
+    private boolean addVerifyPlugin;
+
+    @Getter
+    @Setter
+    private boolean addDirectPlugin;
+
 }